User-API-OTP/requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asgiref==3.6.0
async-generator==1.10
asyncpg==0.27.0
atlastk==0.13.2
attrs==23.1.0
autopep8==2.0.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
Django==4.1.5
django-bootstrap3==22.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
fastapi-login==1.9.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
idna==3.4
install==1.3.5
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.1.1
orjson==3.8.10
outcome==1.2.0
passlib==1.7.4
Pillow==9.4.0
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pypika-tortoise==0.1.6
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==1.0.0
router==0.1
slugify==0.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
starlette==0.26.1
text-unidecode==1.3
tortoise==0.1.1
tortoise-orm==0.19.3
trio==0.21.0
typing_extensions==4.5.0
tzdata==2022.7
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
